refactor(connections): move OAuth refresh rules
