+++ src/treg/domain/connections/__init__.py
@@ -1,0 +1,1 @@
+"""Connection credential rules."""
